# src/main/resources/api-specs/product-category-api-specs.yml
openapi: 3.0.0
info:
  title: Cosmic Cats Marketplace
  description: Marketplace for cats with *cosmic* food<3
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'

paths:
  /api/v1/products:
    get:
      summary: Get all Products
      tags:
        - Product
      responses:
        '200':
          description: List of all products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create new product
      tags:
        - Product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '201':
          description: Successfully created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/products/{id}:
    get:
      summary: Get product by id
      tags:
        - Product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product id
      responses:
        '200':
          description: Found product with id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Partially update a product
      tags:
        - Product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '200':
          description: Successfully updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete product by id
      tags:
        - Product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product id
      responses:
        '204':
          description: No Content — product deleted
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/products/{productId}/{categoryId}:
    patch:
      summary: Add category to product
      tags:
        - Product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product id
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Category id
      responses:
        '200':
          description: Category added to product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product or Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/category:
    get:
      summary: Get all Categories
      tags:
        - Category
      responses:
        '200':
          description: List of all categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryList'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create new category
      tags:
        - Category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
      responses:
        '201':
          description: Successfully created category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/category/{id}:
    get:
      summary: Get category by id
      tags:
        - Category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Category id
      responses:
        '200':
          description: Found category with id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update category
      tags:
        - Category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Category id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
      responses:
        '200':
          description: Successfully updated category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete category by id
      tags:
        - Category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Category id
      responses:
        '204':
          description: No Content — category deleted
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ProductDto:
      type: object
      required:
        - name
        - price
      properties:
        id:
          type: string
          format: uuid
          description: Id of a product
        name:
          type: string
          maxLength: 50
          minLength: 3
          description: Name of a product
        price:
          type: number
          minimum: 1
          description: Price of a product
        category:
          nullable: true
          $ref: '#/components/schemas/CategoryDto'

    ProductList:
      type: object
      properties:
        ProductEntries:
          type: array
          items:
            $ref: '#/components/schemas/ProductDto'

    CategoryDto:
      type: object
      required:
        - name
      properties:
        id:
          type: string
          format: uuid
          description: Id of category
        name:
          type: string
          minLength: 3
          maxLength: 50
          description: Name of a category

    CategoryList:
      type: object
      properties:
        CategoryEntries:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'

    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          description: URI reference identifying the error type
        title:
          type: string
          description: A short description of the error
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Detailed description of the error
        instance:
          type: string
          description: URI of the request instance that caused the error